+++ pilot/pkg/proxy/envoy/v1/watcher_test.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package envoy
+package v1
 
 import (
 	"bytes"
